+++ src/style/Table.scss
@@ -39,7 +39,7 @@
 	margin-right: $gap-10;
 }
 
-.table__ico {
+.table__data__ico {
 	margin-right: $gap-10;
 }
